permission

+++ video/views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import render, get_object_or_404, redirect
 from django.forms import modelformset_factory
 from django.contrib import messages
@@ -33,17 +34,23 @@ def create_video(request):
 
 def update_video(request, id):
     video = get_object_or_404(Video, id=id)
-    video_form = UpdateVideoForm(request.POST or None, request.FILES or None, instance=video)
-    if video_form.is_valid():
-        video = video_form.save()
-        return redirect(video.get_absolute_url())
-    return render(request, 'update_video.html', locals())
+    if request.user == video.user:
+        video_form = UpdateVideoForm(request.POST or None, request.FILES or None, instance=video)
+        if video_form.is_valid():
+            video = video_form.save()
+            return redirect(video.get_absolute_url())
+        return render(request, 'update_video.html', locals())
+    else:
+        return HttpResponse('<h1>403 Forbidden</h1>')
 
 
 def detele_video(request, pk):
     video = get_object_or_404(Video, pk=pk)
-    if request.method == 'POST':
-        video.delete()
-        messages.add_message(request, messages.SUCCESS, 'Successfully deleted!')
-        return redirect("video_list_url")
-    return render(request, 'delete_video.html', locals())
+    if request.user == video.user:
+        if request.method == 'POST':
+            video.delete()
+            messages.add_message(request, messages.SUCCESS, 'Successfully deleted!')
+            return redirect("video_list_url")
+        return render(request, 'delete_video.html', locals())
+    else:
+        return HttpResponse('<h1>403 Forbidden</h1>')
